Refactor insulation logic to support WebMCP and fix radiant barrier calculation
