+++ insulationexplorer.html
@@ -120,6 +120,7 @@
 
     <!-- Logic Engine -->
     <script type="module" src="webmcp.js"></script>
+    <script src="insulation_calculator.js"></script>
     <script>
         // Global Data
         let MATERIALS = {};
@@ -240,86 +241,6 @@
             render();
         }
 
-        // The Physics Engine
-        function calculate(currentStack = stack) {
-            let totalR_winter = 0;
-            let totalR_summer = 0;
-            let warnings = [];
-            let notes = [];
-            
-            // Basic air films (Exterior + Interior)
-            totalR_winter += 0.85; 
-            totalR_summer += 0.85; // simplified
-
-            // Scan stack for context
-            currentStack.forEach((layer, index) => {
-                let r_up = layer.r; // Heat flow up (Winter)
-                let r_down = layer.r; // Heat flow down (Summer)
-                let layerNote = "";
-
-                // --- LOGIC: Radiant Barriers ---
-                if (layer.type === 'radiant') {
-                    // Check neighbors for air gap
-                    const prev = currentStack[index - 1];
-                    const next = currentStack[index + 1];
-                    const hasAirGap = (prev && prev.type === 'air') || (next && next.type === 'air');
-
-                    if (hasAirGap) {
-                        // Radiant Barrier ACTIVE
-                        // Values approximate ASHRAE fundamentals for bright foil w/ air space
-                        
-                        // Heat Flow DOWN (Summer): Radiant barriers excel here
-                        r_down = 10.0; // Effective R-value of the assembly boost
-                        
-                        // Heat Flow UP (Winter): Convection dominates, radiation less factor
-                        r_up = 5.0; 
-                        
-                        layerNote = "ACTIVE (Reflecting Heat)";
-                    } else {
-                        // Radiant Barrier INACTIVE (Conduction only)
-                        r_down = layer.r;
-                        r_up = layer.r;
-                        warnings.push(`Radiant layer at position ${index + 1} has no air gap! It is acting as a conductor only.`);
-                        layerNote = "INACTIVE (No Air Gap)";
-                    }
-                }
-
-                // --- LOGIC: Air Gaps ---
-                if (layer.type === 'air') {
-                    // Air gap R-value depends on facing surfaces. 
-                    // If next to radiant, the radiant layer claims the bonus, keep air nominal
-                    // to avoid double counting.
-                    r_up = 0.9;
-                    r_down = 0.9;
-                }
-
-                // --- LOGIC: Framing Factor (Thermal Bridging) ---
-                // If we have insulation AND structure in the stack, we are strictly adding them linearly here
-                // But in reality, studs bridge the insulation.
-                // For this simple explorer, we will treat "Structure" layers as defining the cavity depth
-                // If a user adds "2x6 Stud" AND "Rockwool", they are occupying the same space visually in this stacker
-                // but physically they merge.
-                // TO FIX: We will treat the stack as a literal sequence through the center of the bay, 
-                // but warn if thickness doesn't match.
-                
-                totalR_winter += r_up;
-                totalR_summer += r_down;
-            });
-
-            // Logic: Check thickness conflicts
-            // E.g. 2x6 stud (5.5") + Rockwool (5.5") + Reflectix + Air Gap
-            // If the user adds framing, we assume they are building the cavity. 
-            // If total insulation thickness > structure thickness, warn.
-            const structureLayer = currentStack.find(l => l.type === 'structure');
-            const insulationThickness = currentStack.filter(l => l.type === 'insulation' || l.type === 'air').reduce((sum, l) => sum + l.thickness, 0);
-            
-            if (structureLayer && insulationThickness > structureLayer.thickness) {
-                warnings.push(`<b>Compression Warning:</b> Your insulation + air gaps (${insulationThickness}") are thicker than your framing (${structureLayer.thickness}"). Compressing insulation reduces R-value and eliminates air gaps.`);
-            }
-
-            return { r_winter: totalR_winter.toFixed(1), r_summer: totalR_summer.toFixed(1), warnings, notes };
-        }
-
         // WebMCP Tool Registration
         // Note: webmcp.js is loaded as a module, so this script might run before it initializes.
         // We can wait for it or just check if it's there.
@@ -401,7 +322,7 @@
                 container.appendChild(div);
             });
 
-            const results = calculate();
+            const results = calculate(stack);
             
             document.getElementById('r-value-winter').innerText = results.r_winter;
             document.getElementById('r-value-summer').innerText = results.r_summer;
